==> code_us/01A_Query_ERA5_Land_CDSAPI_Batch_v2.R <==
# Created by: Zach Popp
# Date Created: 04/01/2025
# Version Number: v2
# Date Modified: 07/08/2025
# Modifications:
#   Switched queue download to separate script
#
# Overview:
#     This code uses the ecmwfr R package to download ERA5 temperature measures
#     from The Copernicus Climate Data Store. More details on the ERA5 API
#     access using the ecmwfr R package are accessible at:
#         https://github.com/bluegreen-labs/ecmwfr
#     Please explore the vignettes provided and note the provided instructions
#     for how to sign up for a Copernicus account and to access the relevant
#     user and key inputs below required to query data.
#
##### Brief Overview of Account Creation: 
##### 1. Create an ECMWF account by self-registering. Follow https://github.com/bluegreen-labs/ecmwfr?tab=readme-ov-file
##### under section "Use: ECMWF Data Store services".
##### 2. Visit user profile to get personal access token. Follow https://github.com/bluegreen-labs/ecmwfr?tab=readme-ov-file
##### under section "Use: ECMWF Data Store services".
##### 3. Visit user profile to accept the terms and conditions in the profile page. 
##### 4. Go to ERA-5 land page following https://cds-beta.climate.copernicus.eu/datasets/reanalysis-era5-land?tab=overview 
##### under section "Data Requests". Go to "Terms of use" block to accept the data licence to use products.
##### 5. Visit user profile page again to double check that Dataset licences to use Copernicus products shows up there and 
##### has been accepted. 
#
#     The Copernicus Climate Data Store API restricts the number of requests
#     actively running at a given time. Requests will sit in a queue, which
#     can be monitored at https://cds.climate.copernicus.eu/requests?tab=all
#
#     Smaller requests move more quickly through the queue. In order to 
#     maximize the request efficiency this request sets up separate requests
#     for the region of interest for each month/year and variable. Based 
#     on information shared through the CDS web forum, the number of hours
#     and number of variables will affect queue time, but the size of the area
#     typically will not. Depending on your time series and area of interest,
#     you may be required to divide your request due to limits on the maximum
#     download size for a single request. 
#
#     This script uses the wf_request_batch function from the ecmwfr package
#     to submit all months of data for a single year and variable at once.
#     This will lead to all requests being added to the queue in short sequence.
#     The batches are submitted in a loop to allow for the addition of more
#     requests to the queue at a given time without waiting for requests to be
#     completed. 
#
#     This script is set up to run via a shell script so that it does not need
#     to be actively monitored in R Studio. A text file with output is written
#     to a Text_Output directory, and this text file is then used in script
#     1C to identify the links where submitted/completed jobs can be 
#     programmatically queried from. Because the ecmwfr requires a user key
#     to be provided, this script is set up to read encrypted text files where
#     API credentials are stored from a directory. It is not recommended to save
#     your API credentials in scripts given that they are sensitive information.
#
#     This example uses the Northeast United States as a test case. The shapefile
#     for the extent is downloaded using the tigris package
#
# Load required packages
#
library("ecmwfr")
library("sf")
library("dplyr")
library("keyring") # Note: the keyring package may not be accessible in your computing
# environment. The package is used to provide a password that
# is otherwise requested directly during an interactive R 
# R Session. If the package cannot be used, this script can 
# be run in your session and the password provided directly.
library("tigris")

# Check package version numbers
#
if (packageVersion("ecmwfr") < "1.5.0"   | packageVersion("sf") < "1.0.16" ) {
  cat("WARNING: packages are outdated and may result in errors.") }

################### User-Define Parameters #####################################
# Set directory. Establishing this at the start of the script is useful in case
# future adjustments are made to the path. 
#
ecmw_dir <- "RawData/ERA5_Hourly" # Directory where rasters will be output to
home_dir <- ""                    # Directory where API credential are stored
trac_dir <- "Code/Track_R/"       # Directory where request syntax will be saved for download

# Set region. This code is developed for US counties with the region representing
# the four US regions in the contiguous US. These could be updated to reflect
# any subdivision of your area of interest, as needed to divide processing into
# more computationally efficient steps.
#
region_in <- 1

# Read in key from file. See https://github.com/bluegreen-labs/ecmwfr for details
# on accessing your key
#
# NOTE: The file storing your API key should never be shared publicly.
# In this script we store the API key in a text file.  api_key will have your CDS API
# and keyring will have a password that is otherwise set and then requested by
# RStudio when attempting to unlock use of the CDS API. If not running in a
# batch script, these approaches may not be required.
#
api_key <- scan(paste0(home_dir, "api_key.txt"), what = "", nmax = 1, quiet = TRUE)
keyring_pass <- scan(paste0(home_dir, "keyring.txt"), what = "", nmax = 1, quiet = TRUE)

# If you are unable to set up the keyring password in your computing environment,
# then you can set your api key directly in the script using wf_set_key(). 
# Note that bash submission may not be effective with this approach, but the 
# data download can be done directly in your R session. 

# Identify extent for download.
# LOAD Shapefile. This approach involves a US application for the Northeast US,
# downloaded using TIGRIS. If you are conducting a global analysis or have 
# an existing shapefile, the below can be replaced to just set the shapefile_cut
# input:
#     shapefile_cut <- st_read("shapefile_path")
#
shapefile_cut <- tigris::states(year = 2020)

# Subset to Northeast region
#
shapefile_cut <- shapefile_cut[shapefile_cut$REGION == region_in, ]

# Set years to download
#
minyear <- 2000
maxyear <- 2024

################### Build Requests #############################################
# Set key. Note - remove this line if using wf_set_key to enter your api
# key directly
# 
keyring_unlock(keyring = "ecmwfr", password = keyring_pass)

# Assess bounding box. The bounding box represents the coordinates of the 
# extent of the shapefile, and will be used to specify the area we would like
# to query from Copernicus Climate Data Store. The API will allow any bounding 
# parameters; however, values that deviate from the original model grid scale
# will be interpolated onto a new grid. Therefore, it’s recommended that for 
# ERA5-Land (which is 0.1˚ resolution) the bounding coordinates be divisible by 
# 0.1 (e.g., 49.5˚N, -66.8˚E, etc.), and that coordinates for ERA5 be divisible
# by 0.25 (e.g., 49.25˚N, -66.75˚E, etc.)
#
input_bbox <- st_bbox(shapefile_cut)

# Add a small buffer around the bounding box to ensure the whole region 
# is queried, and round the parameters to a 0.1 resolution. A 0.1 resolution
# is applied because the resolution of netCDF ERA5 data is .25x.25
# https://confluence.ecmwf.int/display/CKB/ERA5%3A+What+is+the+spatial+reference
#
input_bbox$xmin <- round(input_bbox$xmin[[1]], digits = 1) - 0.1
input_bbox$ymin <- round(input_bbox$ymin[[1]], digits = 1) - 0.1
input_bbox$xmax <- round(input_bbox$xmax[[1]], digits = 1) + 0.1
input_bbox$ymax <- round(input_bbox$ymax[[1]], digits = 1) + 0.1

# The set of inputs below specify the range of years to request, and sets
# the series of month state/end dates to query. There is 
# a limit on the data size that can be downloaded in a given request and smaller
# requests move more quickly through the CDS API queue, so we set up month-
# and variable-level requests in an effort to increase the efficiency
#
query_years <- c(minyear:maxyear)

# Adjust the input years based on the time period for which you want
# to query data
#
query_starts <- c("01-01", "02-01", "03-01", "04-01", "05-01", "06-01", "07-01", 
                  "08-01", "09-01", "10-01", "11-01", "12-01")
query_ends <- c("01-31", "02-29", "03-31", "04-30", "05-31", "06-30", "07-31", 
                "08-31", "09-30", "10-31", "11-30", "12-31")

# Set variables list
#
all_vars <- c("2m_temperature",
              "2m_dewpoint_temperature",
              "surface_net_solar_radiation",
              "surface_net_thermal_radiation",
              "surface_solar_radiation_downwards",
              "surface_thermal_radiation_downwards",
              "10m_u_component_of_wind",
              "10m_v_component_of_wind",
              "surface_pressure")

# Initialize empty lists for each variable
#
combined_request_list <- list()

# To build the requests, we will loop through each year, each variable, and
# each month. We initialize the empty list above and then created nested lists
# within for each year and variable. Separate month-level requests are nested
# in each year-variable sublist. The wf_request_batch function used below
# will pass each request in the list to the API to be filled.
#
for (yr in query_years) {
  
  # Track progress in loop
  #
  cat("Now processing year ", yr, "\n")
  
  # Loop through variables
  #
  for (var in c(all_vars)) {
    
    # Set up year - var index
    #
    yearvar <- paste0(var, yr)
    
    # Initialize the nested list for the variable and year input
    #
    combined_request_list[[as.character(yearvar)]] <- list() 
    
    # Track progress in nested loop
    #
    cat(yearvar, "\n")
    
    # For each year-variable, a monthly request is setup. If a request is
    # too large, it will not be accepted by the CDS servers, so this division
    # of requests is useful and will expedite the process.
    #
    for (i in 1:12) {
      
      # Track progress
      #
      cat("Now processing quarter ", i, "\n")
      
      # Extract inputs for start and end based on list of dates at begin and 
      # end of months around quarter
      #
      query_1 <- query_starts[i]
      query_2 <- query_ends[i]
      
      # Establish query for date periods. This formats the date inputs
      # as they need to be formatted.
      #
      query_dates <- paste0(yr, "-", query_1, "/", yr, "-", query_2)
      
      # The below is the formatted API request language. All of the inputs
      # specified below in proper formatting can be identified by forming a 
      # request using the Copernicus CDS point-and-click interface for data
      # requests. https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-land?tab=form
      # Select the variables, timing, and netcdf as the output format, and then 
      # select "Show API Request" at the bottom of the screen. 
      #
      # Note that the target is the filename that will be exported to the path
      # specified in the next part of the script. If using a loop, ensure that
      # that the unique features of each request are noted in the output! Here
      # we have each of the year, variable, and months (all our loop parameters)
      # in the filename so we won't accidentally overwrite.
      #
      request_era <- list(
        dataset_short_name = "reanalysis-era5-land",
        product_type = "reanalysis",
        variable = c(var),
        date = query_dates,
        time = c('00:00', '01:00', '02:00',
                 '03:00', '04:00', '05:00',
                 '06:00', '07:00', '08:00',
                 '09:00', '10:00', '11:00',
                 '12:00', '13:00', '14:00',
                 '15:00', '16:00', '17:00',
                 '18:00', '19:00', '20:00',
                 '21:00', '22:00', '23:00'),
        data_format = "netcdf",
        download_format = "unarchived",
        area = c(input_bbox$ymax, input_bbox$xmin, input_bbox$ymin, input_bbox$xmax),
        target = paste0("era5-9km-country-", yearvar, "_", query_1, "_", query_2,"_", region_in, ".nc")
      )
      
      # Add request to nested batch list
      #
      combined_request_list[[as.character(yearvar)]] <- c(combined_request_list[[as.character(yearvar)]], list(request_era))   
      
    }
  }
}

################### Submit Requests ############################################
# If you run summary(combined_request_list), you will see that you now
# should have a series of lists of length 12 (1 request for each monht)
# and denoted by a variable name concatenated against a year
# The loop below will pass each list to the wf_request_batch function.
# 
# Within the loop we use 'sink' to ensure all warnings and output are written
# to a text file. These text files will include text that can be uses to reference
# submitted and completed jobs, so that we can programmatically download the data.
#
# The loop below should directly download to your directory, but the text with
# request tracking parameters should still be written to a file as a failsafe.
# Using this loop for direct download can add to the download time as jobs will
# not be added to the queue during the process of files being written to your
# environment. To avoid this delay, you can specify in the path statement a
# directory that does not exist as below:
#
#     wf_request_batch(combined_request_list[[yearvar]], user = "ecmwfr", 
#        path = paste0(ecmw_dir, "/x/"), workers = 12, time_out = 100000)
#
# Adding the /x/ will lead to files failing to be written to your specified 
# environment, but the batch request will continue adding jobs to the queue. 
# We can then add the /x/ directory to the ERA5_Hourly directly before
# running script 01C, so that the download is successful. If you download
# files directly with this loop, script 01C can likely be ignored, just confirm
# that all data is present in your directory.
#
# Loop through years to submit batch script
#
for (year in c(minyear:maxyear)) {
  for (var in c(all_vars)) {
    
    # Form yearvar, which we need to index list. One request list should
    # be submitted at a time
    # 
    yearvar <- paste0(var, year)
    
    # Set up text connection to track processing. This will save the warnings
    # and output to a file on your device, which can then be read in to loop
    # and download.
    #
    log_conn <- textConnection("log_output", "w", local = TRUE)
    sink(log_conn, type = "output")
    sink(log_conn, type = "message")
    
    # We run the submission within a tryCatch function as an error will be 
    # generated by the fake file path. This will only come up after all
    # requests are in
    #
    tryCatch(
      
      # Run the API request - direct download, add /x/ as described above if
      # planning to attempt expedited download.
      batch_submit <- wf_request_batch(combined_request_list[[yearvar]], user = "ecmwfr", 
                                       path = paste0(ecmw_dir), workers = 12, time_out = 100000)
      
      ,
      
      # If there is an error due to a file not being written to your environment,
      # we want the loop to keep going as the file can be recovered using script
      # 1C.
      error = function(e) e
    ) 
    
    # Stop capturing output to file
    #
    sink(NULL, type = "message")
    sink(NULL, type = "output")
    close(log_conn)
    
    # Update processing tracker
    #
    cat(yearvar, " done \n")
    
    # Save text warnings to file
    #
    writeLines(log_output, paste0(trac_dir, "console_test_", yearvar, "_", region_in, "_var.txt"))
    
  }
}
